Move cancelSyncInSyncFramework to SyncFrameworkIntegration
